# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Inspira backend
  version: 0.2.0
  description: |
    HTTP API for the Inspira planning product (repo codename:
    planning-studio). Checked-in copy of the spec FastAPI publishes at
    `/openapi.json` when the app is running. Route shapes are sourced
    from `services/planning_studio_service/api.py` and
    `services/planning_studio_service/auth.py`.
servers:
  - url: http://127.0.0.1:4174
    description: Local dev default
  - url: https://api.tryinspira.com
    description: Production (example)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: inspira_session
      description: |
        Signed itsdangerous cookie set by `/api/auth/signup` or
        `/api/auth/login`. Attributes: HttpOnly, SameSite=Lax,
        Max-Age=30d, Path=/, Secure when INSPIRA_COOKIE_SECURE=true.
...
  schemas:
...
    V2Project:
      type: object
      required:
        - project_id
        - user_id
        - title
        - created_at
        - updated_at
      properties:
        project_id:
          type: string
        user_id:
          type: string
        title:
          type: string
          maxLength: 200
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # -----------------------------------------------------------------
    # Request bodies (v2)
    # -----------------------------------------------------------------
...
    ProjectCreateBody:
      type: object
      properties:
        title:
          type: string
          default: ""
...
    ListProjectsResponse:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/V2Project'
    CreateProjectResponse:
      type: object
      required: [project]
      properties:
        project:
          $ref: '#/components/schemas/V2Project'
...
    LegacyProject:
      type: object
      properties:
        project_id:
          type: string
        title:
          type: string
        summary:
          type: string
        stage:
          type: string
        owner:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      deprecated: true
...
paths:
...
  /api/v2/projects:
    get:
      tags: [v2]
      summary: List active projects for the signed-in user
      operationId: v2ListUserProjects
      security:
        - sessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
    post:
      tags: [v2]
      summary: Create a new v2 project
      operationId: v2CreateProject
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateBody'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
...
  /api/projects:
    get:
      tags: [v1]
      deprecated: true
      summary: List legacy v1 projects
      operationId: v1ListProjects
      security:
        - sessionCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/LegacyProject'
...